- feat : remove lyrics when searching - feat : fix search(slow) before UI render - replace bubble with lava lamp and improve performance
