+++ src/manifest.json
@@ -1,7 +1,7 @@
 {
 	"manifest_version": 3,
 	"name": "Youtube Freemium",
-	"version": "2.3.5",
+	"version": "2.4.0",
 	"description": "Enable lyrics for youtube music videos",
 	"author": "uncannyRishabh",
 	"host_permissions": [
